+++ src/main/java/es/ujaen/dae/indicenciasurbanas/excepciones/IncidenciaEnCurso.java
@@ -1,0 +1,7 @@
+package es.ujaen.dae.indicenciasurbanas.excepciones;
+
+public class IncidenciaEnCurso extends RuntimeException {
+    public IncidenciaEnCurso() {
+
+    }
+}